code documentation
added explanatory documentation
This is a bit non-trivial / non-obvious, even as my own code it took me a while to circle back to this. I’m not a huge fan of these dependencies on Fragment and Activity LifeCycle, but as long as we have required components that aren’t compatible with Compose (aka PresenceViewer) we are stuck with this mixed bag.