Fix occasional reappearance of room when archiving
The issue of occasional flickering and reappearance of a private room
in the rooms list after swiping to archive was because stop() was not
invoked on the observerLocal instance, created when a private rooms was
joined by QR code. This caused its callback closure to fire for every
update on the room instance, including when it was archived, albeit
occasionally because of thread race.