Generalize authentication and add sync to Big Peer
This demo chat app was conceived and implemented as an offline, small peer mesh
chat application. This commit modifies a branch that was especially configured to
authenticate with the Big Peer for testing a specific Big Peer release vesion. This
commit implements conventional Big Peer authentication by a small peer.

The use case for this implementation of the chat app with Big Peer sync should be
considered carefully, as now, what was originally a local small peer mesh by
design is essentially a wide area mesh with a potentially global footprint.
Consequently, careful consideration should be given to managing the eviction of
message and attachment data to avoid overwhelming small peer devices.